flask/mixed/requirements.txt
Authlib
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
SQLAlchemy
